some reorganization for libdg (WIP) + kernel process (WIP)

+++ kernel/cpu/k_defs.h
@@ -148,10 +148,10 @@ struct k_cpu_seg_desc_t
 
 #define K_CPU_SEG_DESC(base, limit, seg_type, dpl, gran, op_size, present) \
     ((struct k_cpu_seg_desc_t){ \
-        .v.w.w0 = ((limit)), \
-        .v.w.w1 = ((base)), \
-        .v.w.w2 = ((((uint32_t)(base)) >> 16) & 0xff) | (seg_type) | (K_CPU_SEG_DESC_DPL(dpl)) | ((uint16_t)(present) << 15), \
-        .v.w.w3 = ((((uint32_t)(limit)) >> 16) & 0xf) | (op_size) | (gran) | ((((uint32_t)(base)) >> 16) & 0xff00)})
+        .v.w.w0 = (uint16_t)(limit), \
+        .v.w.w1 = (uint16_t)(base), \
+        .v.w.w2 = (uint16_t)((((uint32_t)(base)) >> 16) & 0xff) | (uint16_t)(seg_type) | (K_CPU_SEG_DESC_DPL(dpl)) | (uint16_t)((present) << 15), \
+        .v.w.w3 = (uint16_t)((((uint32_t)(limit)) >> 16) & 0xf) | (uint16_t)(op_size) | (uint16_t)(gran) | (uint16_t)((((uint32_t)(base)) >> 16) & 0xff00)})
 
 enum K_CPU_SEG_SEL_FLAGS
 {
